+++ test/unit/account-link-test.js
@@ -3,7 +3,7 @@ var linkGen = require('../../ui/lib/account-link')
 
 describe('account-link', function() {
 
-  it('adds ropsten prefix to ropsten test network', function() {
+  it('adds ropsten prefix to Consortium Test Network', function() {
     var result = linkGen('account', '3')
     assert.notEqual(result.indexOf('ropsten'), -1, 'ropsten included')
     assert.notEqual(result.indexOf('account'), -1, 'account included')