+++ ui/app/components/account-info-link.js
@@ -14,7 +14,7 @@ function AccountInfoLink () {
 
 AccountInfoLink.prototype.render = function () {
   const { selected, network } = this.props
-  const title = 'View account on Etherscan'
+  const title = 'View account on Explorer'
   const url = genAccountLink(selected, network)
 
   if (!url) {