+++ CHANGELOG.md
@@ -124,7 +124,7 @@
 
 ## 2.13.9 2016-11-21
 
-- Add support for the new, default Ropsten Test Network.
+- Add support for the new, default Consortium Test Network.
 - Fix bug that would cause MetaMask to occasionally lose its StreamProvider connection and drop requests.
 - Fix bug that would cause the Custom RPC menu item to not appear when Localhost 8545 was selected.
 - Point ropsten faucet button to actual faucet.